Move generator to .Generator project
