+++ ArgumentativeFilters.sln
@@ -19,7 +19,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Examples", "Examples", "{B9
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExampleMinimalApi", "examples\ExampleMinimalApi\ExampleMinimalApi.csproj", "{AE02EA4A-D6B4-46E6-A12D-9A1857451AA7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ArgumentativeFilters", "src\ArgumentativeFilters\ArgumentativeFilters.csproj", "{C97A8A82-00CB-4DEE-A986-D8503B9A158D}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ArgumentativeFilters.Generator", "src\ArgumentativeFilters.Generator\ArgumentativeFilters.Generator.csproj", "{C97A8A82-00CB-4DEE-A986-D8503B9A158D}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{313FDE40-F228-4BDA-A8FA-E27780961940}"
 EndProject
@@ -41,15 +41,6 @@ Global
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{AE02EA4A-D6B4-46E6-A12D-9A1857451AA7} = {B908D152-98BD-4814-B6F1-AFF36DBD3D62}
-		{634FBCAE-C3C0-4C55-8CCA-6BCCB960CDD6} = {D1D6CAE9-D15C-4172-8658-55B23B1CD63C}
-		{9BA03978-CA26-4082-BFC1-8363983C9324} = {313FDE40-F228-4BDA-A8FA-E27780961940}
-		{0D7E6F0D-1CE5-46D5-AFED-6A974B45CAD5} = {D1D6CAE9-D15C-4172-8658-55B23B1CD63C}
-	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{AE02EA4A-D6B4-46E6-A12D-9A1857451AA7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{AE02EA4A-D6B4-46E6-A12D-9A1857451AA7}.Debug|Any CPU.Build.0 = Debug|Any CPU
@@ -63,5 +54,14 @@ Global
 		{9BA03978-CA26-4082-BFC1-8363983C9324}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9BA03978-CA26-4082-BFC1-8363983C9324}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9BA03978-CA26-4082-BFC1-8363983C9324}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{AE02EA4A-D6B4-46E6-A12D-9A1857451AA7} = {B908D152-98BD-4814-B6F1-AFF36DBD3D62}
+		{634FBCAE-C3C0-4C55-8CCA-6BCCB960CDD6} = {D1D6CAE9-D15C-4172-8658-55B23B1CD63C}
+		{9BA03978-CA26-4082-BFC1-8363983C9324} = {313FDE40-F228-4BDA-A8FA-E27780961940}
+		{0D7E6F0D-1CE5-46D5-AFED-6A974B45CAD5} = {D1D6CAE9-D15C-4172-8658-55B23B1CD63C}
 	EndGlobalSection
 EndGlobal
