+++ src/ArgumentativeFilters.Generator/CodeGeneration/Parameters/Abstract/IFactoryCodeProvider.cs
@@ -1,0 +1,6 @@
+﻿namespace ArgumentativeFilters.Generator.CodeGeneration.Parameters.Abstract;
+
+public interface IFactoryCodeProvider
+{
+    string FactoryCode { get; }
+}
