+++ src/ArgumentativeFilters.Generator/CodeGeneration/Parameters/Abstract/IFilterCodeProvider.cs
@@ -1,0 +1,6 @@
+﻿namespace ArgumentativeFilters.Generator.CodeGeneration.Parameters.Abstract;
+
+public interface IFilterCodeProvider
+{
+    string FilterCode { get; }
+}
