+++ src/ArgumentativeFilters.Generator/CodeGeneration/Parameters/ValueArgumentFilterParameter.cs
@@ -1,14 +1,14 @@
-﻿using ArgumentativeFilters.CodeGeneration.Parameters.Abstract;
-using ArgumentativeFilters.Extensions;
+﻿using ArgumentativeFilters.Generator.CodeGeneration.Parameters.Abstract;
+using ArgumentativeFilters.Generator.Extensions;
 
-namespace ArgumentativeFilters.CodeGeneration.Parameters;
+namespace ArgumentativeFilters.Generator.CodeGeneration.Parameters;
 
 public sealed class ValueArgumentFilterParameter : IndexArgumentFilterParameter, IFilterCodeProvider
 {
     private readonly ITypeSymbol _argumentType;
     private const string ValueNameSuffix = "Value";
-    
-    public ValueArgumentFilterParameter(string argumentName, ITypeSymbol argumentType, bool required) 
+
+    public ValueArgumentFilterParameter(string argumentName, ITypeSymbol argumentType, bool required)
         : base(argumentName, required)
     {
         _argumentType = argumentType;
@@ -20,7 +20,7 @@ public sealed class ValueArgumentFilterParameter : IndexArgumentFilterParameter,
         {
             string typeText = _argumentType.GetFullyQualifiedTypeName();
             string nullableAnnotation = _argumentType.IsReferenceType ? "?" : string.Empty;
-            
+
             return _required
                 ? $"{typeText} {_argumentName}{ValueNameSuffix} = {VariableNames.InvocationFilterContext}.GetArgument<{typeText}>({_argumentName}{IndexNameSuffix}.Value);"
                 : $"{typeText}{nullableAnnotation} {_argumentName}{ValueNameSuffix} = {_argumentName}{IndexNameSuffix}.HasValue ? {VariableNames.InvocationFilterContext}.GetArgument<{typeText}>({_argumentName}{IndexNameSuffix}.Value) : null;";
