+++ src/ArgumentativeFilters.Generator/Parsing/ParameterCodeProviderFactory.cs
@@ -1,16 +1,17 @@
-﻿using ArgumentativeFilters.CodeGeneration.Parameters;
-using ArgumentativeFilters.Extensions;
+﻿using ArgumentativeFilters.Generator.CodeGeneration.Parameters;
+using ArgumentativeFilters.Generator.Extensions;
+
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace ArgumentativeFilters.Parsing;
+namespace ArgumentativeFilters.Generator.Parsing;
 
 public static class ParameterCodeProviderFactory
 {
 
     private const string AspNetCoreEndpointFilterDelegateName = "Microsoft.AspNetCore.Http.EndpointFilterDelegate";
-    
+
     private const string AspNetCoreEndpointFilterInvocationContextName = "Microsoft.AspNetCore.Http.EndpointFilterInvocationContext";
-    
+
     public static ArgumentativeFilterParameterProvider GetParameterCodeProvider(ParameterSyntax parameterSyntax, Compilation compilation)
     {
         var semanticModel = compilation.GetSemanticModel(parameterSyntax.SyntaxTree);
@@ -19,25 +20,25 @@ public static class ParameterCodeProviderFactory
         {
             throw new ApplicationException("failed");
         }
-        
+
         if (parameterSymbol.Type?.ToDisplayString() == AspNetCoreEndpointFilterInvocationContextName)
             return new EndpointFilterInvocationContextFilterParameter();
-        
+
         if (parameterSymbol.Type?.ToDisplayString() == AspNetCoreEndpointFilterDelegateName)
             return new EndpointFilterDelegateFilterParameter();
-        
+
         foreach (var attributeSymbol in parameterSymbol.GetAttributes())
         {
-            if(attributeSymbol?.AttributeClass is null) continue;
-        
+            if (attributeSymbol?.AttributeClass is null) continue;
+
             INamedTypeSymbol attributeContainingTypeSymbol = attributeSymbol.AttributeClass;
             string fullName = attributeContainingTypeSymbol.ToDisplayString();
-        
+
             if (fullName == "ArgumentativeFilters.IndexOfAttribute")
             {
                 return new IndexArgumentFilterParameter(attributeSymbol.ConstructorArguments.FirstOrDefault().Value as string ?? "Invalid IndexOfAttribute parameter.", parameterSymbol.IsParameterRequired());
             }
-            
+
             if (fullName == "Microsoft.AspNetCore.Mvc.FromServicesAttribute")
             {
                 return new ServiceArgumentFilterParameter(parameterSyntax.Identifier.Text, parameterSymbol.Type.GetFullyQualifiedTypeName(), parameterSymbol.IsServiceRequired());
