--- src/ArgumentativeFilters/CodeGeneration/Parameters/Abstract/IFactoryCodeProvider.cs
@@ -1,6 +1,0 @@
-﻿namespace ArgumentativeFilters.CodeGeneration.Parameters.Abstract;
-
-public interface IFactoryCodeProvider
-{
-    string FactoryCode { get; }
-}
