--- src/ArgumentativeFilters/CodeGeneration/Parameters/Abstract/IFilterCodeProvider.cs
@@ -1,6 +1,0 @@
-﻿namespace ArgumentativeFilters.CodeGeneration.Parameters.Abstract;
-
-public interface IFilterCodeProvider
-{
-    string FilterCode { get; }
-}
