+++ tests/ArgumentativeFilters.Functional.Tests/ArgumentativeFilters.Functional.Tests.csproj
@@ -7,7 +7,7 @@
     <NoWarn>CA1034;CA1062;CA1707;CA2007;</NoWarn>
   </PropertyGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\src\ArgumentativeFilters\ArgumentativeFilters.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="true" />
+    <ProjectReference Include="..\..\src\ArgumentativeFilters\ArgumentativeFilters.Generator.csproj" />
   </ItemGroup>
   <ItemGroup>
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
